Added .css() method
A .css() method to get or set css properties of or to element(s). While getting a value, just pass the property name. It will return the computed style for the given property for the first matched element. When setting a property, give it a property and a value argument. It will set them for all the matched elements.