Update docs for wrap plugin
Previously this documentation section included examples which wrap a
`button` element inside an `a` element. This is not permitted in the
[HTML spec](https://w3c.github.io/html-reference/button.button.html#button.button-constraints),
as interactive elements cannot be children of other interactive
elements.

This was also an inaccessible example, seeing as typical keyboard and
screenreader focus for buttons and links will be either inconsistent or
broken depending on the browser

Went through and updated the examples to change the `button` to a
`span`, which is proper accessible HTML 👍